@@ -9,14 +9,14 @@ use Tests\TestCase;
 
 class RegistrationTest extends TestCase
 {
-    public function testRegistrationScreenCanBeRendered(): void
+    public function test_registration_screen_can_be_rendered(): void
     {
         $response = $this->get('/registrieren');
 
         $response->assertStatus(200);
     }
 
-    public function testNewUsersCanRegister(): void
+    public function test_new_users_can_register(): void
     {
         $response = $this->post('/registrieren', [
             'name' => 'Test User',
