@@ -13,7 +13,7 @@ class InstrumentResourceTest extends TestCase
         $this->setupAdmin();
     }
 
-    public function testRenderingResourcePage()
+    public function test_rendering_resource_page()
     {
 
         $this->get(InstrumentResource::getUrl())
