@@ -13,7 +13,7 @@ class SongResourceTest extends TestCase
         $this->setupAdmin();
     }
 
-    public function testRenderingResourcePage()
+    public function test_rendering_resource_page()
     {
 
         $this->get(SongResource::getUrl())
