@@ -15,13 +15,13 @@ class EmailsTest extends TestCase
         $this->actingAs($this->createUserForInternalRoutes());
     }
 
-    public function testRouteReturnsSuccessful()
+    public function test_route_returns_successful()
     {
         $this->get('intern/emails')
             ->assertSuccessful();
     }
 
-    public function testCorrectView()
+    public function test_correct_view()
     {
         $this->get('intern/emails')
             ->assertInertia(
