@@ -7,13 +7,13 @@ use Tests\TestCase;
 
 class HomeTest extends TestCase
 {
-    public function testRouteReturnsSuccessful()
+    public function test_route_returns_successful()
     {
         $this->get('/')
             ->assertSuccessful();
     }
 
-    public function testCorrectView()
+    public function test_correct_view()
     {
         $this->get('/')
             ->assertInertia(
